+++ packages/electron/src/main/builder/BacktraceClientBuilder.ts
@@ -1,10 +1,14 @@
-import { BacktraceClientBuilder as NodeBacktraceClientBuilder, BacktraceNodeClientSetup } from '@backtrace/node';
+import { BacktraceNodeClientSetup, BacktraceClientBuilder as NodeBacktraceClientBuilder } from '@backtrace/node';
+import { BacktraceClient } from '../BacktraceClient';
+import { AllWindowsAttributeProvider } from '../attributes/AllWindowsAttributeProvider';
 import { AppAttributeProvider } from '../attributes/AppAttributeProvider';
 import { ApplicationInformationAttributeProvider } from '../attributes/ApplicationInformationAttributeProvider';
 import { GpuAttributeProvider } from '../attributes/GpuAttributeProvider';
 import { GpuFeatureAttributeProvider } from '../attributes/GpuFeatureAttributeProvider';
+import { NetAttributeProvider } from '../attributes/NetAttributeProvider';
 import { ReadyAppAttributeProvider } from '../attributes/ReadyAppAttributeProvider';
-import { BacktraceClient } from '../BacktraceClient';
+import { ScreenAttributeProvider } from '../attributes/ScreenAttributeProvider';
+import { WindowEventSubscriber } from '../breadcrumbs/WindowEventSubscriber';
 
 export class BacktraceClientBuilder extends NodeBacktraceClientBuilder {
     constructor(clientSetup: BacktraceNodeClientSetup) {
@@ -15,6 +19,11 @@ export class BacktraceClientBuilder extends NodeBacktraceClientBuilder {
         this.addAttributeProvider(new GpuFeatureAttributeProvider());
         this.addAttributeProvider(new ReadyAppAttributeProvider());
         this.addAttributeProvider(new AppAttributeProvider());
+        this.addAttributeProvider(new AllWindowsAttributeProvider());
+        this.addAttributeProvider(new NetAttributeProvider());
+        this.addAttributeProvider(new ScreenAttributeProvider());
+
+        this.useBreadcrumbSubscriber(new WindowEventSubscriber());
     }
 
     public build(): BacktraceClient {