+++ packages/sdk-core/src/common/DelayHelper.ts
@@ -1,3 +1,5 @@
+import { AbortError } from './AbortError';
+
 export class Delay {
     /**
      * Promise set timeout wrapper.
@@ -6,12 +8,21 @@ export class Delay {
      */
     public static wait(timeout: number, signal?: AbortSignal) {
         return new Promise<void>((resolve, reject) => {
+            // intervalId has to be defined here,
+            // as abortCallback can execute before setTimeout is executed
+            // eslint-disable-next-line prefer-const
+            let intervalId: ReturnType<typeof setTimeout> | undefined;
+
             function abortCallback() {
                 clearTimeout(intervalId);
-                reject(new Error('Operation cancelled.'));
+                reject(new AbortError());
             }
 
-            const intervalId = setTimeout(() => {
+            if (signal?.aborted) {
+                return abortCallback();
+            }
+
+            intervalId = setTimeout(() => {
                 signal?.removeEventListener('abort', abortCallback);
                 resolve();
             }, timeout);